// apps/web/lib/generated/models/file-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * danmaku API
 * Danmaku API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface FileItemDto {
    /**
     * ファイル/フォルダID
     */
    'id': string;
    /**
     * ファイル/フォルダ名
     */
    'name': string;
    /**
     * MIME タイプ - \'application/vnd.google-apps.folder\' = フォルダ - \'video/mp4\' = MP4 ビデオ
     */
    'mimeType': string;
    /**
     * ファイル/フォルダサイズ（バイト） フォルダの場合は null
     */
    'size'?: number;
    /**
     * 最終更新日時（ISO 8601形式）
     */
    'modifiedTime': string;
    /**
     * Google Drive WebView URL
     */
    'webViewLink': string;
    /**
     * サムネイル URL（ビデオファイルの場合のみ）
     */
    'thumbnailLink'?: string;
    /**
     * 親フォルダID
     */
    'parentId'?: string;
}